Remove session start event.
